requirements.txt
albumentations==1.3.1
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.1.0
streamlit==1.29.0
timm==0.9.12
torch
tqdm==4.66.1
transformers
